Add logging for prediction classes for instances
Still incomplete rendition as of this date. The logic behind class '2'
was derived from the COCO car ID. The idea to filter a Tensor was
derived from https://stackoverflow.com/questions/57570043/filter-data-in-pytorch-tensor

TODO  Still need to figure out how to get the correct class's position
in the frame to put the timer at an appropriate position.