Improve prediction mask to include all vehicles
The prediction mask now uses a range of IDs from 2--8, which are all the
vehicle IDs present in the test dataset. The vector in the indices
tensor has also been flattened. Some comments have been added to explain
the code. Finally, more information is being output to the console to
show the frame timestamp and panoptic segmentation information.