Add custom visualizer
Create a custom visualizer based on the VideoVisualizer utility class in
the Detectron2 toolkit. This class draws instance predictions showing
either the mask or boundary box, along with information about the time
that a vehicle has been in frame.

Also added in this commit are small refactorings in the 'predictor.py'
functions.